+++ packages/vendure-plugin-utm-tracker/src/types.ts
@@ -23,7 +23,7 @@ export interface AttributionModel {
   name: string;
   /**
    * Determine the attribution percentage for each UTM parameter of the given order.
-   * The parameters are already sorted by connectedAt timestamp, ascending (oldest first)
+   * The parameters are already sorted by connectedAt timestamp, descending (newest first)
    */
   calculateAttribution(
     utmParameters: UtmOrderParameter[],